requirements.txt
# ============================================
# Bibliothèques Python géospatiales (essentielles)
# ============================================

# --- Core ---
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
pandas>=2.0.0

# --- Géométrie et vecteur ---
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0
geopandas>=0.14.0
pyshp>=2.3.0
geojson>=3.0.0

# --- Raster ---
rasterio>=1.3.0
xarray>=2023.0.0
rioxarray>=0.15.0

# --- PostgreSQL/PostGIS ---
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0

# --- Analyse spatiale ---
rtree>=1.0.0
libpysal>=4.9.0
esda>=2.5.0

# --- Clustering ---
scikit-learn>=1.2.0

# --- Géocodage ---
geopy>=2.3.0

# --- Visualisation ---
matplotlib>=3.7.0
folium>=0.14.0
plotly>=5.18.0

# --- Utilitaires ---
requests>=2.28.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0

# --- Validation ---
jsonschema>=4.17.0
pydantic>=2.5.0